refactor: 숫자가 클 경우Long 타입이 필요할 수 있어 Integer->Long 으로 타입 수정

+++ src/main/java/calculator/calculator/Calculator.java
@@ -5,9 +5,9 @@ import java.util.ArrayList;
 public class Calculator {
 
 
-    public static int allSum(ArrayList<Integer> savedNumberRepository) {
+    public static Long allSum(ArrayList<Long> savedNumberRepository) {
 
-        int sum = 0;
+        long sum = 0;
         for (int i = 0; i < savedNumberRepository.size(); i++) {
             sum += savedNumberRepository.get(i);
         }
